+++ CHANGELOG.md
@@ -13,11 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Collection type must always have an ID field.
 - Changed PHP file representation from an associative array to a strongly-typed FileObject class.
 - Reduced payload size for `Collection.get`.
 - Use modern async/await for client code (supported in all major browers/Node.js since 2017).
 - Made `firestorm.files.get` a generic method for increased type safety.
 - Stricter type checking for `array-contains` search methods.
+- Mixed arrays of strings and numbers are allowed in all bulk Firestorm methods.
 
 ### Fixed